Correctly handle state changes
When the user sets the thermostat as off, turn off recirculation. This will stay off till the next scheduled event.
When the user says they want heat, turn on the schedule override and turn on recirculation for the next 5 mins.
When the user changes to Auto, then return to the current schedule.