Get ready to control a Navien
When a NaviLink unit disappears, check the state of the scheduler and set the Navien to the appropriate state. If the scheduler isn't running, then don't do anything.
The Eve app's Thermo support wants to turn the heat on if the current water temperature is below the set point when a schedule is running. So trick it, and when the schedule says to set the temp to Economy temperature (aka turn off recirculation) tell Eve that the schedule is not running.